fix: certificates overflow on s42.app (#251)
**Describe the pull request**

The certificate of s42.app includes `*.s42.app` and occurs an overflow
on top of `dashboards.s42.app` and `next.s42.app`. Usage of wildcard
with idle sockets / ssl sockets can occurs a 404 when use two
certificates as the same endpoint.

Remove the wildcard certificate to prevent this error. 

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no