feat: reimplement the last connected at based on the location we received (#260)
**Relative Issues:**
https://discord.com/channels/248936708379246593/1019744956535013436

**Describe the pull request**

On v3, no " offline since" is present. This is a missing feature
compared to v2. This PR reimplements the last connected at strategy, but
only on the location we already received. This is to prevent any
overload call chain on the intranet.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no

**Additional context**

**Add a new field `lastLocation` to the model user. Can be maybe replace
the `currentLocation` relation in the future. This is a duplicated
field.**