perf: duo avatar url in 2020+ is too big and create latency on app (#262)
**Describe the pull request**

Since 2 years, the full image url is a 1500x1500 image and when navigate
on app with a lot of big picture create huge latency. Save the small
version under `duoAvatarSmallURL` is an acceptable workaround. In the
future, create a service to resize and cache on the fly to have always
the desired picture size.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no

**Additional context**
This pull Request needs to us to run a manual task to re-import small
version url