fix: the user is not correctly considered as active (#280)
**Relative Issues:** Resolve #279 

**Describe the pull request**
One column on the model user called `is_a_user` must be set as true when
an User takes the action to join the app. But actually, the number of
active users seems to lower. This PR fix this issue

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no

**Additional context**
A manual migration in the database will be perfromed to set all user
with duo account linked as active.