fix: meilisearch tasks is not cleaned and increase the volume size (#357)
**Relative Issues:** Resolve #356 

**Describe the pull request**

Meilsiearch create a task for each async action. When we perform an
update of an object in meilisearch, MS create a task. In 1 week we
create million of tasks and that is never cleaned. This PR implement a
cronjob who delete old tasks to keep volume clean and small as possible.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no