security: change webhooks secrets (#407)
**Describe the pull request**

The purpose of this Pull Request is to change the webhook secrets due to
a potential leak. We have identified a vulnerability in the system that
could allow unauthorized third parties to access these secrets and cause
damage to our system. To avoid any potential risk, we have decided to
change the secrets and implement an additional layer of security to
protect our system.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no

**Additional context**
Special thanks to **clafoutis** to her support