fix: webhooks nack cause storm on rabbitmq, 42api, loki and sentry (#424)
**Relative Issues:** Resolve #408

**Describe the pull request**
This pull request addresses the issue of webhook failures causing a
negative acknowledgment (NACK) storm on RabbitMQ, as well as affecting
intra, Loki, and Sentry. The PR provides a fix that improves the webhook
handling by implementing a dead letter queue, which prevents NACK storms
and reduces the impact on the mentioned services. By utilizing a dead
letter queue to handle failed webhook messages, we can maintain the
application's stability and performance, ensuring a smoother experience
for users and better resource management.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no