fix: initialize sentry for edge runtime (#426)
**Describe the pull request**

This pull request addresses the issue of Sentry not being initialized
for the edge runtime environment. The fix ensures that Sentry is
properly set up to capture and report errors and performance metrics
from the edge runtime. By initializing Sentry for edge runtime, we can
better monitor and troubleshoot issues that may arise in this
environment, leading to an improved overall application experience and
more effective error tracking.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no