feat: user profile as slideshow and animations  (#434)
**Relative Issues:** Resolve #8 

**Describe the pull request**

This Pull Request introduces several enhancements to the user profile,
including the ability to display them as a slideshow with customizable
options. Additionally, we have added animations to the user-related
features, such as the cluster map and profile slideshow, to provide a
more engaging user experience.

**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no

**Additional context**

These changes will improve the user experience and make the user
profiles more visually appealing and customizable. The addition of
animations to the user-related features will also make the overall user
experience more enjoyable and engaging. We introduce `framer-motion` on
the project to test it, to ensure that they work as intended and do not
impact the performance or stability of the system.