feat: api cache system with dragonfly (#436)
**Describe the pull request**
This Pull Request introduces a new feature to our API - a cache system
implemented with Dragonfly. The cache system is primarily used to cache
all proxies from the intranet, aiming to prevent bursts on the intra
API. By caching frequently requested data, the system reduces the load
on the intra API, resulting in improved performance and response time.

Dragonfly, a robust caching library, is utilized to handle the caching
mechanism effectively, providing features such as cache expiration and
invalidation. The cache system is designed to gracefully handle cache
expiration and ensure that the most up-to-date data is served to the
users.

**Checklist**

- [ ] I have linked the relative issue to this pull request
- [ ] I have made the modifications or added tests related to my PR
- [ ] I have added/updated the documentation for my RP
- [ ] I put my PR in Ready for Review only when all the checklist is
checked

**Breaking changes ?**
no