fix: upload to s3 in ovh differ than s3 specs (#451)
**Describe the pull request**
This pull request addresses an issue where the upload process to S3 in
OVH differs from the standard S3 specifications. The fix involves
aligning the upload process with the standard S3 protocol, ensuring
compatibility and smooth operation. This fix allow user to upload and
access in public-read aby cover uploaded

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP