fix: latency and white theme flickering caused by next-auth issues  (#457)
**Describe the pull request**
This pull request addresses an issue causing latency and white theme
flickering due to problems with next-auth. The fix involves resolving
these next-auth issues by rollback to client side session fetching and
eliminate the undesirable flickering effect by rolling back the client
cache on local storage.

**Checklist**
- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP

**Additional context**
The latency seams to be when nextauth try to store new session on
response cookies. Maybe is a next issue. We put some metadata to the
transactions for a better monitoring on Sentry