feat: celebrating pride month 🏳️‍🌈 (#458)
**Describe the pull request**
This pull request introduces a new feature in celebration of Pride
Month. This feature aims to show support and recognition for the LGBTQ+
community, aligning with the spirit of inclusivity and diversity that
Pride Month represents. The implementation details of this feature can
range from visual changes in the app's theme to added content that
celebrates and honors the LGBTQ+ community. By introducing this feature,
we aim to create a more inclusive environment for all our users,
demonstrating our commitment to diversity and respect for all.

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP

**Additional context**
The pronoun settings were added prior to this pull request, signifying
the initial step towards inclusivity. Be yourself and no longer fear
being true to who you are. 🌈 🏳️‍🌈