refactor(auth)!: rework all the auth service to remove grpc integrally (#459)
**Describe the pull request**
This pull request undertakes a significant overhaul of the
authentication service, the main objective being to eliminate the gRPC
integration entirely. In the process of this major refactoring, we have
successfully removed the sandbox stack, enabling entirely offline
development. The intended outcome of these changes is to create a more
efficient, maintainable, and scalable authentication service. Please
note that this is a breaking change and may lead to incompatibility with
earlier versions or configurations of the application.

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP

**Additional context**
We've also reworked the authentication stack to eliminate calls to
next-auth, leading to enhanced performance.