feat: add personal workspace entity on cluster map (#463)
**Describe the pull request**
This pull request introduces a new entity that adds a personal workspace
entity on the cluster map. The addition of personal workspaces on the
map provides users with a visual representation of their personal
workspace within the cluster. This addition aims to improve user
interaction and understanding of their workspace distribution in the
cluster, allowing users to easily identify where they can work with
their personal computers.

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP