chore: migrate clusters map for paris campus to noc42 (#462)
**Describe the pull request**

This pull request involves the migration of the clusters map for the Paris campus to noc42. The reason for this migration is due to construction work at 42 Paris, where the old building no longer exists, and has been replaced by noc42. This task ensures that the clusters map accurately represents the current state of the campus, providing an up-to-date resource for users. By completing this migration, we maintain the accuracy and relevance of our campus information, ensuring a smooth transition to the new noc42 setting.

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP

---------

Signed-off-by: Atomys <contact@atomys.fr>
Co-authored-by: Atomys <contact@atomys.fr>