chore: upgrade monitoring stack major version (#470)
**Describe the pull request**
This pull request involves a major version upgrade of our monitoring
stack. The goal of this task is to keep our application up-to-date with
the latest enhancements, bug fixes, and potential performance
improvements offered by this new version. By updating to the latest
version of our monitoring stack, we aim to ensure the optimal
performance, reliability, and feature set of our monitoring
capabilities, ensuring we can effectively track and manage the health
and performance of our application.

**Checklist**

- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP