fix: paris campus locations regexp (#478)
**Describe the pull request**
This pull request fixes an issue where the link from the friends page on the Paris campus was invalid. The problem stemmed from a combination of factors: a regular expression (regexp) that didn't accommodate for the possibility of uppercase letters in the location. The fix involves adjusting the regular expression to correctly handle locations with uppercase letters, thereby rectifying the invalid link issue. By resolving this problem, we ensure that all links within our application are valid and functional, providing a seamless user experience.

**Checklist**
- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP

---------

Co-authored-by: Atomys <contact@atomys.fr>