revert: postgres upgrade to version 15 rollbacked to 14.8 (#497)
**Describe the pull request**
This pull request involves reverting a previous upgrade of our
PostgreSQL system from version 14.8 to 15. After careful evaluation and
in response to observed issues post-upgrade, we have decided to rollback
to version 14.8 to maintain the stability and reliability of our system.
While we always strive to keep our technologies up-to-date, our priority
is to provide a consistent and dependable service. We will continue to
assess the suitability of PostgreSQL 15 and consider future upgrades
when appropriate.

**Checklist**
- [x] I have made the modifications or added tests related to my PR
- [x] I have run the tests and linters locally and they pass
- [x] I have added/updated the documentation for my RP