feat: implement various quality of life on cluster maps (#512)
**Describe the pull request**

This PR introduces a series of quality of life improvements to the
cluster maps. These enhancements aim to make the maps more
user-friendly, intuitive, and efficient, providing a better user
experience for navigation and overall interaction with the campus map
features. Like:

1. **devexp**: The `totalWorkspaces` in the cluster map definition is no
longer required. This value is now calculated from the given map.
2. Overriding a workspace with no connected user will provide you with
the workspace identifier.
3. When a cluster doesn't have a name, don't display the identifier
twice.
4. The sidebar no longer sticks when scrolling over the cluster sidebar.
5. The scrollbar is now styled to match the application's design.

**Checklist**

- [ ] I have made the modifications or added tests related to my PR
- [ ] I have run the tests and linters locally and they pass
- [ ] I have added/updated the documentation for my RP